+++ api/helpers/things.js
@@ -1,5 +1,8 @@
 let { isString } = require("lodash");
 let log = require("winston");
+const cache = require("apicache");
+const equals = require("deep-equal");
+const moment = require("moment");
 
 const { as, db, sql } = require("./db");
 const { getUserIfExists } = require("../helpers/user");
@@ -13,7 +16,6 @@ function addRelatedList(owner_type, owner_id, related_type, id_list) {
     id_list = [id_list];
   }
   owner_id = as.number(owner_id);
-  // case and related come from our code, we'll trust those
   let values = id_list
     .map(id => {
       let escaped_id = as.number(id);
@@ -38,7 +40,6 @@ function removeRelatedList(owner_type, owner_id, related_type, id_list) {
     id_list = [id_list];
   }
   owner_id = as.number(owner_id);
-  // case and related come from our code, we'll trust those
   return id_list
     .map(id => {
       let escaped_id = as.number(id);
@@ -69,13 +70,7 @@ function diffRelatedList(first, second) {
 }
 
 function getXByIdFns(type) {
-  let retFns = {};
-  const typeUC = type[0].toUpperCase() + type.slice(1);
-  retFns[`get${typeUC}ById_lang_userId`] = async function(
-    thingId,
-    lang,
-    userId
-  ) {
+  const getById_lang_userId = async function(thingId, lang, userId) {
     const theThing = await db.one(sql(`../sql/${type}_by_id.sql`), {
       thingId,
       lang
@@ -89,20 +84,21 @@ function getXByIdFns(type) {
     return theThing;
   };
 
-  retFns[`get${typeUC}ByRequest`] = async function(req) {
+  const getByRequest = async function(req) {
     const thingId = as.number(req.params[`${type}Id`]);
     const lang = as.value(req.params.language || "en");
     const userId = await getUserIfExists(req);
-    return await retFns[`get${typeUC}ById_lang_userId`](thingId, lang, userId);
+    return await getById_lang_userId(thingId, lang, userId);
   };
 
-  retFns[`return${typeUC}ById`] = async function(req, res) {
+  const returnById = async function(req, res) {
     try {
-      const thing = await retFns[`get${typeUC}ByRequest`](req);
+      const thing = await getByRequest(req);
       res.status(200).json({ OK: true, data: thing });
     } catch (error) {
       log.error(
-        "Exception in GET /case/%s => %s",
+        "Exception in GET /%s/%s => %s",
+        type,
         req.params[`${type}Id`],
         error
       );
@@ -112,7 +108,209 @@ function getXByIdFns(type) {
       });
     }
   };
-  return retFns;
+  return {
+    getById_lang_userId,
+    getByRequest,
+    returnById
+  };
+}
+
+const getByType_id = {
+  case: getXByIdFns("case"),
+  method: getXByIdFns("method"),
+  organization: getXByIdFns("organization")
+};
+
+function getEditXById(type) {
+  return async function editById(req, res) {
+    cache.clear();
+    const thingId = as.number(req.params[type + "Id"]);
+    try {
+      // FIXME: Figure out how to get all of this done as one transaction
+      const lang = as.value(req.params.language || "en");
+      const userId = req.user.user_id;
+      const oldThing = await getByType_id[type].getById_lang_userId(
+        thingId,
+        lang,
+        userId
+      );
+      const newThing = req.body;
+      let updatedText = {
+        body: oldThing.body,
+        title: oldThing.title,
+        language: lang,
+        type: type,
+        id: thingId
+      };
+      let updatedThingFields = [];
+      let isTextUpdated = false;
+      let anyChanges = false;
+      let retThing = null;
+
+      /* DO ALL THE DIFFS */
+      Object.keys(oldThing).forEach(async key => {
+        if (
+          // All the ways to check if a value has not changed
+          newThing[key] === undefined ||
+          equals(oldThing[key], newThing[key]) ||
+          (/_date/.test(key) &&
+            moment(oldThing[key]).format() ===
+              moment(newThing[key]).format()) ||
+          (/related_/.test(key) &&
+            equals(
+              oldThing[key].map(x => x.id),
+              newThing[key].map(x => x.id || x.value)
+            ))
+        ) {
+          // skip, do nothing, no change for this key
+          if (key === "communication_mode") {
+            console.log(
+              "skipping %s: %s = %s",
+              key,
+              oldThing[key],
+              newThing[key]
+            );
+          } else {
+            console.log("skipping %s", key);
+          }
+        } else if (!equals(oldThing[key], newThing[key])) {
+          anyChanges = true;
+          // If the body or title have changed: add a record in X__localized_texts
+          if (key === "body" || key === "title") {
+            updatedText[key] = newThing[key];
+            isTextUpdated = true;
+            // If related_cases, related_methods, or related_organizations have changed
+            // update records in related_nouns
+          } else if (
+            [
+              "related_cases",
+              "related_methods",
+              "related_organizations"
+            ].includes(key)
+          ) {
+            // DELETE / INSERT any needed rows for related_nouns
+            const oldList = oldThing[key];
+            const newList = newThing[key];
+            newList.forEach(x => x.id = x.id || x.value); // handle client returning value vs. id
+            const diff = diffRelatedList(oldList, newList);
+            const relType = key.split("_")[1].slice(0, -1); // related_Xs => X
+            const add = addRelatedList(
+              type,
+              thingId,
+              relType,
+              diff.add.map(x => x.id)
+            );
+            const remove = removeRelatedList(
+              type,
+              thingId,
+              relType,
+              diff.remove.map(x => x.id)
+            );
+            if (add || remove) {
+              await db.none(add + remove);
+            }
+            anyChanges = true;
+            // If any of the fields of thing itself have changed, update record in appropriate table
+          } else if (["id", "post_date", "updated_date"].includes(key)) {
+            console.warn(
+              "Trying to update a field users shouldn't update: %s",
+              key
+            );
+            // take no action
+          } else if (key === "featured" && !user.groups.includes("Curators")) {
+            console.warn("Non-curator trying to update Featured flag");
+            // take no action
+          } else if (key === "location") {
+            updatedThingFields.push({
+              key: as.name(key),
+              value: as.location(newThing[key])
+            });
+          } else if (key === "lead_image") {
+            var img = newThing[key];
+            updatedThingFields.push({
+              key: as.name(key),
+              value: as.attachment(img.url, img.title, img.size)
+            });
+          } else if (["other_images", "files"].includes(key)) {
+            updatedThingFields.push({
+              key: as.name(key),
+              value: as.attachments(newThing[key])
+            });
+          } else {
+            console.log(
+              "Unspecified change found in %s: %s != %s",
+              key,
+              oldThing[key],
+              newThing[key]
+            );
+            let value = oldThing[key];
+            let asValue = as.text;
+            if (typeof value === "boolean") {
+              asValue = as.value;
+            } else if (value === null) {
+              value = "null";
+              asValue = as.value;
+            } else if (typeof value === "number") {
+              asValue = as.number;
+            }
+            updatedThingFields.push({
+              key: as.name(key),
+              value: asValue(value)
+            });
+          }
+        }
+      }); // end of for loop over object keys
+      if (anyChanges) {
+        // Actually make the changes
+        if (isTextUpdated) {
+          // INSERT new text row
+          await db.none(sql("../sql/insert_localized_text.sql"), updatedText);
+        }
+        // Update last_updated
+        updatedThingFields.push({ key: "updated_date", value: as.text("now") });
+        // UPDATE the thing row
+        // console.log(
+        //   "QUERY >>>%s<<<",
+        //   as.format(sql("../sql/update_noun.sql"), {
+        //     keyvalues: updatedThingFields
+        //       .map(field => field.key + " = " + field.value)
+        //       .join(", "),
+        //     type: type,
+        //     id: thingId
+        //   })
+        // );
+        await db.none(sql("../sql/update_noun.sql"), {
+          keyvalues: updatedThingFields
+            .map(field => field.key + " = " + field.value)
+            .join(", "),
+          type: type,
+          id: thingId
+        });
+        // INSERT row for X__authors
+        await db.none(sql("../sql/insert_author.sql"), {
+          user_id: userId,
+          type: type,
+          id: thingId
+        });
+        // update materialized view for search
+        retThing = await getByType_id[type].getById_lang_userId(
+          as.number(thingId),
+          lang,
+          userId
+        );
+      } else {
+        // end if anyChanges
+        retThing = oldThing;
+      } // end if not anyChanges
+      res.status(200).json({ OK: true, data: retThing });
+    } catch (error) {
+      log.error("Exception in PUT /%s/%s => %s", type, thingId, error);
+      res.status(500).json({
+        OK: false,
+        error: error
+      });
+    } // end catch
+  };
 }
 
 module.exports = {
@@ -120,5 +318,7 @@ module.exports = {
   removeRelatedList,
   getXByIdFns,
   diffRelatedList,
-  difference
+  difference,
+  getEditXById,
+  getByType_id
 };
